make the fetchcontracts domain model work
