+++ ledger-service/http-json/src/main/scala/com/digitalasset/http/util/ClientUtil.scala
@@ -1,5 +1,6 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
+
 package com.daml.http.util
 
 import java.util.UUID
@@ -8,7 +9,7 @@ import com.daml.ledger.api.refinements.ApiTypes.{CommandId, Party}
 import com.daml.ledger.api.v1.transaction_filter.{Filters, TransactionFilter}
 import com.daml.ledger.api.{v1 => lav1}
 
-object ClientUtil {
+private[http] object ClientUtil {
   def uniqueId(): String = UUID.randomUUID.toString
 
   def uniqueCommandId(): CommandId = CommandId(uniqueId())
@@ -17,5 +18,5 @@ object ClientUtil {
     TransactionFilter(Party.unsubst(ps).map((_, Filters.defaultInstance)).toMap)
 
   def boxedRecord(a: lav1.value.Record): lav1.value.Value =
-    lav1.value.Value(lav1.value.Value.Sum.Record(a))
+    com.daml.fetchcontracts.util.boxedRecord(a)
 }
